fix(cli): address metrics review gates
