+++ contracts/functional-scenarios.json
@@ -177,6 +177,16 @@
       "reviewedReason": "No qualifying repository-owned functional test currently invokes and observes this component through its customer boundary."
     },
     {
+      "stableId": "cli/you.metrics",
+      "name": "you metrics",
+      "interface": "cli",
+      "classification": "runnable",
+      "status": "missing",
+      "lane": "short",
+      "executionClass": "deterministic",
+      "reviewedReason": "Focused CLI behavior is covered by package tests, but no qualifying repository-owned functional test currently invokes and observes this component through its customer boundary."
+    },
+    {
       "stableId": "cli/you.models",
       "name": "you models",
       "interface": "cli",
